requirements.txt
streamlit==1.28.2
pymupdf==1.23.8
pytesseract==0.3.10
pillow==10.1.0
opencv-python-headless==4.8.1.78
langdetect==1.0.9
dateparser==1.1.8
regex==2023.10.3
fpdf==1.7.2
openai==1.3.5
ollama==0.1.7
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1